move option to folder
